+++ example/lib/widgets/keys.dart
@@ -1,9 +1,0 @@
-/* open_ui
- * Copyright (c) 2022-2024 Empathetech LLC. All rights reserved.
- * See LICENSE for distribution and usage details.
- */
-
-import 'package:flutter/material.dart';
-
-final GlobalKey<ScaffoldMessengerState> scaffoldMessengerKey =
-    GlobalKey<ScaffoldMessengerState>();