+++ lib/empathetech_flutter_ui.dart
@@ -98,5 +98,6 @@ export 'src/widgets/user_customization/dominant_hand_switch.dart';
 export 'src/widgets/user_customization/image_setting.dart';
 export 'src/widgets/user_customization/layout_setting.dart';
 export 'src/widgets/user_customization/locale_setting.dart';
+export 'src/widgets/user_customization/random_button.dart';
 export 'src/widgets/user_customization/reset_button.dart';
 export 'src/widgets/user_customization/theme_mode_switch.dart';