+++ lib/src/consts/consts.dart
@@ -26,5 +26,8 @@ const double largeBreakpoint = 1000.0;
 /// 0.08
 const double highlightOpacity = 0.08;
 
-/// 0.4 == [highlightOpacity] * 5
-const double containerOpacity = 0.4;
+/// 0.40
+const double selectionOpacity = 0.40;
+
+/// 0.25
+const double containerOpacity = 0.25;