+++ lib/src/functions/dialogs.dart
@@ -34,18 +34,6 @@ Future<void> logAlert(
   );
 }
 
-/// Log the passed message and display a [SnackBar] to notify the user
-Future<void> logSnack(
-  ScaffoldMessengerState state,
-  String message,
-) async {
-  debugPrint(message);
-  state.showSnackBar(SnackBar(
-    content: Text(message, textAlign: TextAlign.center),
-    duration: readingTime(message),
-  ));
-}
-
 /// Wrap a [ColorPicker] in an [EzAlertDialog]
 Future<dynamic> ezColorPicker(
   BuildContext context, {